联动删除

+++ database/migrations/2020_05_28_130843_create_topics_table.php
@@ -7,11 +7,11 @@ class CreateTopicsTable extends Migration
 {
 	public function up()
 	{
-		Schema::create('topics', function(Blueprint $table) {
+   	    Schema::create('topics', function(Blueprint $table) {
             $table->increments('id');
             $table->string('title')->index();
             $table->text('body');
-            $table->bigInteger('user_id')->unsigned()->index();
+            $table->integer('user_id')->unsigned()->index();
             $table->integer('category_id')->unsigned()->index();
             $table->integer('reply_count')->unsigned()->default(0);
             $table->integer('view_count')->unsigned()->default(0);
